Start using Interp as varname for TclInterp
It's been a Tcl convention approximately forever that a local C variable
containing an interpreter pointer (Tcl_Interp *) is named interp and
I've followed that with the UnsafeMutablePointer<Tcl_Interp> equivalent.

I had also been calling instances of the Swift wrapper class TclInterp
"interp" but have changed that to "Interp" to make it more clear the
difference / avoid confusion.